Actually fixes button overlap in 1.19.4 to 1.20.2
In versions 1.19.3 to 1.20.2, the Fabric API overrides the "render"
method in PressableWidget instead of the "renderButton" or
"renderWidget" methods, which results in the buttons being set to render
and be active every frame, resulting in the previous solution I did
being completely ineffective.

This was resolved in later versions of the Fabric API, but I've been
forced to add a hacky workaround to disable the button rendering for
versions 1.19.4, 1.20.1, and 1.20.2.